+++ server/edusync/src/main/java/com/codestates/edusync/auth/userdetails/MemberDetailsService.java
@@ -53,6 +53,8 @@ public class MemberDetailsService implements UserDetailsService {
             return getEmail();
         }
 
+
+
         @Override
         public boolean isAccountNonExpired() {
             return true;
