+++ server/edusync/src/main/java/com/codestates/edusync/study/plancalendar/dto/TimeScheduleSingleResponseDto.java
@@ -1,4 +1,4 @@
-package com.codestates.edusync.infodto.timeschedule.dto;
+package com.codestates.edusync.study.plancalendar.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Getter;
@@ -10,10 +10,12 @@ import java.sql.Timestamp;
 @NoArgsConstructor
 @Getter
 @Setter
-public class TimeScheduleInfoDto {
-    @JsonFormat(pattern = "MM-dd HH:mm", timezone = "Asia/Seoul")
-    private Timestamp start;
+public class TimeScheduleSingleResponseDto {
+    private Long id;
 
     @JsonFormat(pattern = "MM-dd HH:mm", timezone = "Asia/Seoul")
-    private Timestamp end;
+    private Timestamp startTime;
+
+    @JsonFormat(pattern = "MM-dd HH:mm", timezone = "Asia/Seoul")
+    private Timestamp endTime;
 }
