+++ server/edusync/src/main/java/com/codestates/edusync/study/plancalendar/studygroup/repository/CalendarStudygroupRepository.java
@@ -1,6 +1,6 @@
 package com.codestates.edusync.study.plancalendar.studygroup.repository;
 
-import com.codestates.edusync.infodto.timeschedule.entity.TimeSchedule;
+import com.codestates.edusync.study.plancalendar.entity.TimeSchedule;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
