+++ server/edusync/src/main/java/com/codestates/edusync/study/studygroupjoin/controller/StudygroupJoinController.java
@@ -35,8 +35,8 @@ public class StudygroupJoinController {
 
         List<StudygroupJoin> studygroupJoinList;
 
-        if (join) studygroupJoinList = studygroupJoinService.findStudygroupJoinList(studygroupId); // 멤버 리스트
-        else studygroupJoinList = studygroupJoinService.findStudygroupJoinCandidateList(studygroupId); // 대기 리스트
+        if (join)   studygroupJoinList = studygroupJoinService.findStudygroupJoinList(studygroupId); // 멤버 리스트
+        else        studygroupJoinList = studygroupJoinService.findStudygroupJoinCandidateList(studygroupId); // 대기 리스트
 
         StudygroupJoinDto.Response studygroupJoinDtos =
                 studygroupJoinmapper.studygroupJoinToStudygroupJoinDtos(studygroupJoinList);
