+++ server/edusync/src/main/java/com/codestates/edusync/study/studygroupjoin/service/StudygroupJoinService.java
@@ -9,6 +9,7 @@ import com.codestates.edusync.study.studygroup.service.StudygroupService;
 import com.codestates.edusync.study.studygroupjoin.entity.StudygroupJoin;
 import com.codestates.edusync.study.studygroupjoin.repository.StudygroupJoinRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
